Analyze the build_node field
A field needs to be analyzed in order to be searched using wildcards,
and the only field that contains information about the type of nodepool
node or the cloud region used is build_node. So analyzing this field
will allow us to use searches such as build_node:"devstack-precise*" or
build_node:"*hpcloud*" compounded with an error message to track errors
across devstack-precise nodes or in hpcloud respectively.

Change-Id: Ie1a4cc55fdc0e858060f215526abfb52137a1061