Add config options for cors config
With newer ES we need to set CORS controls properly because it seems to
care about that now. This change allows you to enable the use of CORS
and specify which origins are allowed.

By default ES doesn't allow any cross origin requests so you have to
explicitly enable it with these options if you want to allow it.

Change-Id: I0aa8d5167c770c1024b7596da582d6cc089b1b47