Add elasticsearch mapping for @timestamp field
Instead of storing the @timestamp as another string, store it in 'date'
format. Also store it as a doc_value, which is stored on-disk rather
than in-memory. Elasticsearch 2.0 makes doc_value the default, but we are
currently using 1.7.5.

Adding this mapping will not alter current data, it will only affect
data indexed into new indices after the mapping is in place.

This change is motivated by indices.breaker warnings I am seeing for the
@timestamp field. There is one node in my cluster that is failing to
load the timestamps into the heap. After inspecting
'localhost:9200/_nodes/stats/indices/fielddata?fields=*&pretty' I saw
that @timestamp was using an order of magnitude more heap than any other
field.

Change-Id: I7c8c8f7ea032f8a93ab8306de2c97352e4b4c899